Farm batch_norm serialization tests off into a separate file
